ai-engine/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pandas==1.3.3
numpy==1.21.2
requests==2.26.0
tensorflow==2.6.0
scikit-learn==0.24.2
matplotlib==3.4.3
folium==0.12.1
pillow==8.3.2
transformers==4.11.3
torch==1.9.1
nltk==3.6.3